Corregido package y clase principal para despliegue en Render

+++ src/main/java/com/example/ciudades/dto/CiudadDto.java
@@ -2,22 +2,22 @@ package com.example.ciudades.dto;
 
 public class CiudadDto {
 
-    private Long id;
+    private int id;
     private String nombre;
 
     public CiudadDto() {
     }
 
-    public CiudadDto(Long id, String nombre) {
+    public CiudadDto(int id, String nombre) {
         this.id = id;
         this.nombre = nombre;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
